+++ src/views/About/HistoryTab.tsx
@@ -1,0 +1,19 @@
+import React from 'react'
+
+import {
+    AppBar,
+    Tab,
+    Tabs,
+    Typography,
+    Box,
+    Paper,
+    Grid,
+    List,
+    ListItemText,
+  } from "@mui/material";
+
+export const HistoryTab = () => {
+  return (
+    <div>HistoryTab</div>
+  )
+}
