+++ src/views/About/InfoTab.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const InfoTab = () => {
+  return (
+    <div>InfoTab</div>
+  )
+}
